requirements.txt
django==1.11
djangorestframework
django-cors-headers
mysqlclient
Pillow
requests
web3
coreapi

